pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "navsim"
version = "0.1.0"
description = "Minimal 2D robot navigation demo with planning and control."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "matplotlib",
  "imageio",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
navsim-demo = "navsim.cli:main"
navsim-benchmark = "navsim.benchmark:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["navsim*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
